feat(eth-contracts): add v2 tests
